+++ src/Components/Expression.php
@@ -220,7 +220,8 @@ final class Expression implements Component
 
             // Skipping whitespaces and comments.
             if (($token->type === Token::TYPE_WHITESPACE) || ($token->type === Token::TYPE_COMMENT)) {
-                if ($isExpr) {
+                // If the token is a closing C comment from a MySQL command, it must be ignored.
+                if ($isExpr && $token->token !== '*/') {
                     $ret->expr .= $token->token;
                 }