+++ src/Token.php
@@ -109,6 +109,23 @@ class Token
      */
     public const TYPE_LABEL = 10;
 
+    /**
+     *  All tokens types
+     */
+    public const TYPE_ALL = [
+        self::TYPE_NONE,
+        self::TYPE_KEYWORD,
+        self::TYPE_OPERATOR,
+        self::TYPE_WHITESPACE,
+        self::TYPE_COMMENT,
+        self::TYPE_BOOL,
+        self::TYPE_NUMBER,
+        self::TYPE_STRING,
+        self::TYPE_SYMBOL,
+        self::TYPE_DELIMITER,
+        self::TYPE_LABEL,
+    ];
+
     // Flags that describe the tokens in more detail.
     // All keywords must have flag 1 so `Context::isKeyword` method doesn't
     // require strict comparison.
@@ -292,8 +309,8 @@ class Token
 
                 if (isset($str[0]) && (($str[0] === '`') || ($str[0] === '"') || ($str[0] === '\''))) {
                     $quote = $str[0];
-                    $str = str_replace($quote . $quote, $quote, $str);
                     $str = mb_substr($str, 1, -1, 'UTF-8');
+                    $str = str_replace($quote . $quote, $quote, $str);
                 }
 
                 return $str;