+++ src/TokensList.php
@@ -7,6 +7,7 @@ namespace PhpMyAdmin\SqlParser;
 use ArrayAccess;
 
 use function count;
+use function in_array;
 use function is_array;
 use function is_string;
 
@@ -119,7 +120,7 @@ class TokensList implements ArrayAccess
      */
     public function getPrevious(): ?Token
     {
-        for (; $this->idx > 0; --$this->idx) {
+        for (; $this->idx >= 0; --$this->idx) {
             if (
                 ($this->tokens[$this->idx]->type !== Token::TYPE_WHITESPACE)
                 && ($this->tokens[$this->idx]->type !== Token::TYPE_COMMENT)
@@ -132,16 +133,42 @@ class TokensList implements ArrayAccess
     }
 
     /**
+     * Gets the previous token.
+     *
+     * @param int|int[] $type the type
+     *
+     * @return Token|null
+     */
+    public function getPreviousOfType($type)
+    {
+        if (! is_array($type)) {
+            $type = [$type];
+        }
+
+        for (; $this->idx >= 0; --$this->idx) {
+            if (in_array($this->tokens[$this->idx]->type, $type, true)) {
+                return $this->tokens[$this->idx--];
+            }
+        }
+
+        return null;
+    }
+
+    /**
      * Gets the next token.
      *
-     * @param int $type the type
+     * @param int|int[] $type the type
      *
      * @return Token|null
      */
     public function getNextOfType($type)
     {
+        if (! is_array($type)) {
+            $type = [$type];
+        }
+
         for (; $this->idx < $this->count; ++$this->idx) {
-            if ($this->tokens[$this->idx]->type === $type) {
+            if (in_array($this->tokens[$this->idx]->type, $type, true)) {
                 return $this->tokens[$this->idx++];
             }
         }