+++ tests/Components/ExpressionTest.php
@@ -53,7 +53,7 @@ class ExpressionTest extends TestCase
     /**
      * @return string[][]
      */
-    public function parseErrProvider(): array
+    public static function parseErrProvider(): array
     {
         return [
             /*
@@ -91,5 +91,64 @@ class ExpressionTest extends TestCase
             Expression::build($component),
             '1 + 2 AS `three`, 1 + 3 AS `four`'
         );
+    }
+
+    /**
+     * @return string[][]
+     */
+    public static function mysqlCommandsProvider(): array
+    {
+        return [
+            [
+                '/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;',
+                'SET  @OLD_CHARACTER_SET_CLIENT = @@CHARACTER_SET_CLIENT',
+            ],
+            [
+                '/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;',
+                'SET  @OLD_CHARACTER_SET_RESULTS = @@CHARACTER_SET_RESULTS',
+            ],
+            [
+                '/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;',
+                'SET  @OLD_COLLATION_CONNECTION = @@COLLATION_CONNECTION',
+            ],
+            [
+                '/*!40101 SET NAMES utf8 */;',
+                'SET NAMES utf8',
+            ],
+            [
+                '/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;',
+                'SET  @OLD_TIME_ZONE = @@TIME_ZONE',
+            ],
+            [
+                "/*!40103 SET TIME_ZONE='+00:00' */;",
+                "SET  TIME_ZONE = '+00:00'",
+            ],
+            [
+                '/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;',
+                'SET  @OLD_UNIQUE_CHECKS = @@UNIQUE_CHECKS, UNIQUE_CHECKS = 0',
+            ],
+            [
+                '/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;',
+                'SET  @OLD_FOREIGN_KEY_CHECKS = @@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS = 0',
+            ],
+            [
+                "/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;",
+                "SET  @OLD_SQL_MODE = @@SQL_MODE, SQL_MODE = 'NO_AUTO_VALUE_ON_ZERO'",
+            ],
+            [
+                '/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;',
+                'SET  @OLD_SQL_NOTES = @@SQL_NOTES, SQL_NOTES = 0',
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider mysqlCommandsProvider
+     */
+    public function testMysqlCommands(string $expr, string $expected): void
+    {
+        $parser = new Parser($expr, true);
+        $parser->parse();
+        self::assertSame($expected, $parser->statements[0]->build());
     }
 }