+++ tests/Parser/SelectStatementTest.php
@@ -28,7 +28,7 @@ class SelectStatementTest extends TestCase
     /**
      * @return string[][]
      */
-    public function selectProvider(): array
+    public static function selectProvider(): array
     {
         return [
             ['parser/parseSelect2'],
@@ -44,6 +44,8 @@ class SelectStatementTest extends TestCase
             ['parser/parseSelect12'],
             ['parser/parseSelect13'],
             ['parser/parseSelect14'],
+            ['parser/parseSelect15'],
+            ['parser/parseSelect16'],
             ['parser/parseSelectErr1'],
             ['parser/parseSelectErr2'],
             ['parser/parseSelectNested'],
@@ -97,6 +99,9 @@ class SelectStatementTest extends TestCase
             ['parser/parseSelectWithParenthesis'],
             ['parser/parseSelectOrderByComment'],
             ['parser/parseSelectOverAlias_mariadb_100600'],
+            ['parser/parseSelectGroupBy'],
+            ['parser/parseSelectGroupByErr'],
+            ['parser/parseSelectGroupByWithComments'],
         ];
     }
 }