+++ tests/Parser/TransactionStatementTest.php
@@ -4,10 +4,22 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Parser;
 
+use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class TransactionStatementTest extends TestCase
 {
+    public function testBuildWithoutEnd(): void
+    {
+        $data = $this->getData('parser/parseTransaction4');
+        $parser = new Parser($data['query']);
+        $stmt = $parser->statements[0];
+        $this->assertEquals(
+            'START TRANSACTION;SET  time_zone = "+00:00";',
+            $stmt->build()
+        );
+    }
+
     /**
      * @dataProvider transactionProvider
      */
@@ -19,12 +31,13 @@ class TransactionStatementTest extends TestCase
     /**
      * @return string[][]
      */
-    public function transactionProvider(): array
+    public static function transactionProvider(): array
     {
         return [
             ['parser/parseTransaction'],
             ['parser/parseTransaction2'],
             ['parser/parseTransaction3'],
+            ['parser/parseTransaction4'],
             ['parser/parseTransactionErr1'],
         ];
     }