+++ tests/TestCase.php
@@ -11,8 +11,8 @@ use PhpMyAdmin\SqlParser\Lexer;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Token;
 use PhpMyAdmin\SqlParser\TokensList;
+use PhpMyAdmin\SqlParser\Tools\CustomJsonSerializer;
 use PHPUnit\Framework\TestCase as BaseTestCase;
-use Zumba\JsonSerializer\JsonSerializer;
 
 use function file_get_contents;
 use function str_contains;
@@ -96,7 +96,7 @@ abstract class TestCase extends BaseTestCase
         $serializedData = file_get_contents('tests/data/' . $name . '.out');
         $this->assertIsString($serializedData);
 
-        $serializer = new JsonSerializer();
+        $serializer = new CustomJsonSerializer();
         $data = $serializer->unserialize($serializedData);
 
         $this->assertIsArray($data);